+++ scripts/build-es-dictionary.ts
@@ -21,7 +21,9 @@ const LANG = 'es';
 const spellcheckerDictionary1st = fs.readFileSync(`./resources/${LANG}/${DICTIONARIES.spellchecker.dir}/dictionary.txt`, 'utf-8');
 const spellcheckerDictionary2nd = fs.readFileSync(`./resources/${LANG}/${DICTIONARIES.spellcheckerAlt.dir}/dictionary.txt`, 'utf-8');
 const winningDictionary = fs.readFileSync(`./resources/${LANG}/${DICTIONARIES.winning.dir}/dictionary.txt`, 'utf-8');
-const winningDictionaryAlt = fs.readFileSync(`./resources/${LANG}/${DICTIONARIES.winningAlt.dir}/dictionary.txt`, 'utf-8');
+const frequencyDictionary = fs.readFileSync(`./resources/${LANG}/${DICTIONARIES.popularity.dir}/dictionary.txt`, 'utf-8');
+
+const frequencyWords = getWordsFromDictionary(frequencyDictionary, { pattern: 'word ignore', lang: LANG });
 
 const spellcheckerWords1st = getWordsFromDictionary(spellcheckerDictionary1st, { pattern: 'word ignore', lang: LANG });
 const spellcheckerWords2nd = getWordsFromDictionary(spellcheckerDictionary2nd, { pattern: 'word ignore', lang: LANG });
@@ -31,7 +33,7 @@ const spellcheckerWords = [...new Set([...spellcheckerWords1st, ...spellcheckerW
 
 const [longestWinningWords, ...winningWordsDictionariesToCheck] = [
     getWordsFromDictionary(winningDictionary, { pattern: 'word ignore', lang: LANG }),
-    getWordsFromDictionary(winningDictionaryAlt, { pattern: 'word ignore', lang: LANG }),
+    frequencyWords,
 ].sort((a, b) => b.length - a.length);
 
 const winningWords = longestWinningWords.filter(
@@ -50,4 +52,5 @@ actionBuildDictionary(
     },
     spellcheckerWords,
     winningWords,
+    frequencyWords,
 );