+++ src/components/EndResult/EndResult.tsx
@@ -20,10 +20,12 @@ import IconEgg from '@components/Icons/IconEgg';
 import IconGamepad from '@components/Icons/IconGamepad';
 
 import Button from '@components/Button/Button';
-import GoToDictionaryButtons from '@components/Dictionary/GoToDictionaryButtons';
+import GoToDictionaryButton from '@components/Dictionary/GoToDictionaryButton';
 
 import ShareButton from '@components/Share/ShareButton';
 import StatisticsHint from '@components/Panes/Statistics/StatisticsHint';
+
+import WordPopularity from '@features/wordPopularity/components/WordPopularity';
 
 import EndResultGameTime from './EndResultGameTime';
 import EndResultSummary from './EndResultSummary';
@@ -101,7 +103,8 @@ function EndResult() {
               )}
               {gameMode !== GameMode.SandboxLive && <ShareButton shouldShowSettings />}
           </div>
-          <GoToDictionaryButtons words={guesses.map(({ word }) => word)} />
+          <GoToDictionaryButton word={wordToGuess} />
+          <WordPopularity word={wordToGuess} />
           <EndResultGameTime />
           {gameMode === GameMode.Daily && (
           <p